move package.json code to separate file
